refactor(super-editor): extract header/footer variant normalization
Move normalizeVariant out of the doc-api adapter helper into a dedicated
module under presentation-editor/header-footer/, where its sole non-adapter
caller (HeaderFooterSessionManager) lives. The slot-materialization helper
re-exports it for existing import sites.
